eslint改为双引号lint 首页图集id=94

+++ webpage/4_vue-element-admin/.eslintrc.js
@@ -162,7 +162,7 @@ module.exports = {
       }
     }],
     'padded-blocks': [2, 'never'],
-    'quotes': [2, 'single', {
+    'quotes': [2, 'double', {
       'avoidEscape': true,
       'allowTemplateLiterals': true
     }],
